// src/types/HiringRoleApplicantProcess.ts
// }
export interface HiringRoleApplicantProcess {
    id: string,
    hiring_role_applicant_id: string,
    hiring_role_process_id: string,
    hiring_role_process_step_id: string,
    is_admin_approved: boolean,
    video_response: string
    created_by: string
    updated_by: string
    deleted_by: string
    created_at: string
    updated_at: string
}

   //     {
    //   "hiring_role_applicant_id": "string",
    //   "hiring_role_process_id": "string",
    //   "hiring_role_process_step_id": "string",
    //   "is_admin_approved": true,
    //   "video_response": "string"
    // }